chore(🚚): 경로를 변경합니다.
