--- Box42/Main/BoxViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  URLViewModel.swift
-//  Box42
-//
-//  Created by Chanhee Kim on 8/9/23.
-//
-
-import Foundation
-import Combine
-
-// CRUD 4가지 형태의 데이터 가공 create, read, update, delete
-class BoxViewModel: ObservableObject {
-    // Box View Model은 전체적인 View Model을 담당하게 될것.
-}
