--- Box42/Menubar/MenubarView.swift
@@ -1,13 +1,0 @@
-//
-//  MenubarView.swift
-//  Box42
-//
-//  Created by Chan on 2023/03/16.
-//
-
-import Foundation
-import AppKit
-
-class MenuBarView {
-    
-}
